test: reject message (de)serialization
This adds tests for the previously untested reject message
(de)serialization. The two reject messages were received from an
older Bitcoin Core peer that still sends reject messages. The test
fails as the rust-bitcoin (de)serialization is incorrect. The
"message" is treated as CommandString which is (de)serialized as 12
byte string. However, it should be a variable length string. This is
fixed in the next commit.